refactor(dvrp): inline DvrpGlobalTravelTimesMatrixProvider
